Remove `GraknTestEngineSetup`
